Organized header components inside header folder
