+++ src/app/header/search-bar/search-bar.component.scss
