+++ src/app/header/selector/selector.component.scss
