+++ src/app/header/slider/slider.component.scss
