+++ src/app/header/slider/slider.component.spec.ts
