+++ src/app/header/slider/slider.module.ts
