+++ src/app/header/slider/slider.stories.ts
